refactor: message 패키지 변경 및 상세 예외 메시지 값 던지기 추가
