+++ Sep.09/src/main/java/ui/TouchScreen.java
@@ -1,5 +1,7 @@
 package ui;
 
+import common.message.GuidanceMessage;
+
 public class TouchScreen {
     Reader reader;
     Writer writer;
